# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Woreda 11 Club Attendance Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members (admin sees all, manager sees own club)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: clubId
          in: query
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberList"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member details with attendance history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberDetail"
    put:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List published news with read status
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Publish a news article (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: News published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get a single news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    put:
      operationId: updateNews
      tags: [news]
      summary: Update a news article (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated news
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete a news article (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /news/{id}/read:
    post:
      operationId: markNewsRead
      tags: [news]
      summary: Mark a news article as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Member:
      type: object
      required: [id, clubId, fullName, phoneNumber, age, gender, createdAt]
      properties:
        id:
          type: string
        clubId:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [Male, Female, Other]
        addedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        clubName:
          type: ["string", "null"]
    MemberInput:
      type: object
      required: [clubId, fullName, phoneNumber, age, gender]
      properties:
        clubId:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [Male, Female, Other]
    MemberUpdate:
      type: object
      properties:
        fullName:
          type: string
        phoneNumber:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [Male, Female, Other]
    MemberList:
      type: object
      required: [members, total]
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        total:
          type: integer
    AttendanceRecord:
      type: object
      required: [id, memberId, memberName, attendedAt]
      properties:
        id:
          type: string
        memberId:
          type: string
        memberName:
          type: string
        memberAge:
          type: integer
        memberGender:
          type: string
        attendedAt:
          type: string
        sessionDate:
          type: string
        clubName:
          type: ["string", "null"]
    MemberDetail:
      type: object
      required: [id, clubId, fullName, phoneNumber, age, gender, createdAt, attendanceHistory]
      properties:
        id:
          type: string
        clubId:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        age:
          type: integer
        gender:
          type: string
        createdAt:
          type: string
        clubName:
          type: ["string", "null"]
        attendanceHistory:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
...
    NewsItem:
      type: object
      required: [id, title, description, isPublished, publishedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        publishedBy:
          type: ["string", "null"]
        publishedByName:
          type: ["string", "null"]
        isPublished:
          type: boolean
        publishedAt:
          type: string
        updatedAt:
          type: string
        isRead:
          type: boolean
        readCount:
          type: integer
    NewsInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
    NewsUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
...