# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Woreda 11 Club Attendance Management API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberUpdateGender } from './memberUpdateGender';

export interface MemberUpdate {
  fullName?: string;
  phoneNumber?: string;
  age?: number;
  gender?: MemberUpdateGender;
}